Panic: Save Pre-Exception Registers, Memory/IO header for Crash Debug
* added a bl subroutine Panic_SaveRegisters
  + runs assembly instructions to save registers r0-r12 and lr to
  crash_regs[13]

* added header structure to be saved during each memory dump
  + cc_header_t (start, end) meets standards of Crash Debug

* included Exceptions.S in Makefile.platform to run Panic_SaveRegisters
  + ldr pseudocode does not seem to work with inline asm in tinyos
  + bl to the routine in Exceptions.S allows us to run without error

+++ tos/platforms/mm6a/startup.c
@@ -113,7 +113,7 @@ const image_info_t image_info __attribute__ ((section(".image_meta"))) = {
 ow_control_block_t ow_control_block __attribute__ ((section(".overwatch_data")));
 
 /* space to save registers during crash processing */
-uint8_t crash_regs[CRASH_REGS_BYTES]   __attribute__ ((section(".crash_regs")));
+uint32_t crash_regs[CRASH_REGS_BYTES]   __attribute__ ((section(".crash_regs")));
 
 /* Crash stack is used by unhandled Exception/Fault and Panic */
 uint8_t crash_stack[CRASH_STACK_BYTES] __attribute__ ((section(".crash_stack")));
